split utils into own file
