+++ gui/slick/js/utils/get-meta.js
@@ -1,0 +1,5 @@
+const getMeta = pyVar => {
+    return $('meta[data-var="' + pyVar + '"]').data('content');
+};
+
+export default getMeta;
