+++ gui/slick/js/utils/is-meta.js
@@ -1,0 +1,6 @@
+const isMeta = (pyVar, result) => {
+    const reg = new RegExp(result.length > 1 ? result.join('|') : result);
+    return (reg).test($('meta[data-var="' + pyVar + '"]').data('content'));
+};
+
+export default isMeta;
