+++ gui/slick/js/utils/notify-modal.js
@@ -1,0 +1,6 @@
+const notifyModal = message => {
+    $('#site-notification-modal .modal-body').html(message);
+    $('#site-notification-modal').modal();
+};
+
+export default notifyModal;
